requirements.txt
dbt-snowflake>=1.11.0
python-dotenv>=1.0.0
fastapi>=0.135.0
uvicorn[standard]>=0.42.0
httpx>=0.28.0
pydantic-settings>=2.0.0
snowflake-connector-python>=4.2.0
